+++ CHANGELOG.md
@@ -6,4 +6,12 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 ## [Unreleased]
 
-- Initial release
+## [1.0.0] - 2023.09.30
+
+### Added
+
+- All xicons packages support.
+
+### Fixed
+
+- Bad regex perfermance of xicons match.